adminQueries.py combined
